Extract project_core
