+++ crates/project/Cargo.toml
@@ -42,6 +42,7 @@ node_runtime.workspace = true
 parking_lot.workspace = true
 postage.workspace = true
 prettier.workspace = true
+project_core.workspace = true
 rand.workspace = true
 regex.workspace = true
 rpc.workspace = true
@@ -71,6 +72,7 @@ release_channel.workspace = true
 lsp = { workspace = true, features = ["test-support"] }
 prettier = { workspace = true, features = ["test-support"] }
 pretty_assertions.workspace = true
+project_core = { workspace = true, features = ["test-support"] }
 rpc = { workspace = true, features = ["test-support"] }
 settings = { workspace = true, features = ["test-support"] }
 unindent.workspace = true
