+++ crates/project_core/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "project_core"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+test-support = [
+    "client/test-support",
+    "language/test-support",
+    "settings/test-support",
+    "text/test-support",
+    "gpui/test-support",
+]
+
+[dependencies]
+anyhow.workspace = true
+client.workspace = true
+clock.workspace = true
+collections.workspace = true
+fs.workspace = true
+futures.workspace = true
+fuzzy.workspace = true
+git.workspace = true
+gpui.workspace = true
+ignore = "0.4.22"
+itertools.workspace = true
+language.workspace = true
+log.workspace = true
+lsp.workspace = true
+parking_lot.workspace = true
+postage.workspace = true
+rpc.workspace = true
+schemars.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+settings.workspace = true
+smol.workspace = true
+sum_tree.workspace = true
+text.workspace = true
+util.workspace = true
+
+[dev-dependencies]
+clock = {workspace = true, features = ["test-support"]}
+collections = { workspace = true, features = ["test-support"] }
+git2.workspace = true
+gpui = {workspace = true, features = ["test-support"]}
+rand.workspace = true
+settings = {workspace = true, features = ["test-support"]}
+pretty_assertions.workspace = true
