+++ crates/project_core/src/ignore.rs
