+++ crates/project_core/src/project_settings.rs
